+++ src/Configuration/Options.php
@@ -32,6 +32,7 @@ class Options implements OptionsInterface
         'exceptions' => Option\Exceptions::class,
         'prefix' => Option\Prefix::class,
         'crc16' => Option\CRC16::class,
+        'driver_info' => Option\DriverInfo::class,
     ];
 
     /** @var array */